ingestion/pyproject.toml
[project]
name = "hypertrending-ingestion"
version = "0.1.0"
description = "Data ingestion service for HyperTrending - fetches trends from Google Trends and other sources"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytrends>=4.9.2",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "schedule>=1.2.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
hypertrending = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 100
